+++ app/src/main/java/com/misw4203/vinilos/domain/usecase/RemoveAlbumFromCollectorUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RemoveAlbumFromCollectorUseCase @Inject constructor(
     private val repository: CollectorRepository,
 ) {
-    suspend operator fun invoke(collectorId: Int, albumId: Int) =
+    suspend operator fun invoke(collectorId: Int, albumId: Long) =
         repository.removeAlbumFromCollector(collectorId, albumId)
 }
