+++ app/src/main/java/com/misw4203/vinilos/presentation/ui/screens/band/AddAlbumToBandScreen.kt
@@ -146,9 +146,9 @@ fun AddAlbumToBandScreen(
 @Composable
 private fun ScreenContent(
     form: AddAlbumToBandFormState,
-    addingId: Int?,
+    addingId: Long?,
     onQueryChange: (String) -> Unit,
-    onAdd: (Int) -> Unit,
+    onAdd: (Long) -> Unit,
 ) {
     val currentAlbums = form.allAlbums.filter { it.id in form.currentAlbumIds }
 
@@ -201,8 +201,8 @@ private fun ScreenContent(
             items(form.filteredAvailable, key = { it.id }) { album ->
                 AlbumRow(
                     album = album,
-                    isAdding = addingId == album.id.toInt(),
-                    onAdd = { onAdd(album.id.toInt()) },
+                    isAdding = addingId == album.id,
+                    onAdd = { onAdd(album.id) },
                     modifier = Modifier.testTag("available_album_band_${album.id}"),
                 )
                 HorizontalDivider(
